refactor/RLT-56: create PhoneINput as a single component
