+++ src/presentation/components/PhoneInput/README.md
@@ -1,0 +1,52 @@
+# TextInput
+
+O componente é um `input` para entrada de números, já com a estilização do projeto. Ele também pode receber as propriedades do input padrão. Também possui a estilização de quando o campo recebe valores inválidos ou não atende os requisitos.
+
+## 💪 Funcionalidades
+
+- Validação
+- Suporte para inúmeros países
+
+## 🖥️ Screenshots
+
+![InputText Preview](https://i.ibb.co/7KzjNQc/Captura-de-tela-2024-05-01-081643.png)
+
+## 🛠️ Stack utilizada
+
+**Front-end:**
+
+- JavaScript
+- React
+- Styled-Components
+- react-international-phone (input library)
+- google-libphonenumber (input validation library)
+
+## 🟩 Cores
+
+| Tipo                   | Hexadecimal                                                      |
+| ---------------------- | ---------------------------------------------------------------- |
+| input/background       | ![#f3f4f6](https://via.placeholder.com/10/f3f4f6?text=+) #f3f4f6 |
+| input/text-color       | ![#091030](https://via.placeholder.com/10/091030?text=+) #091030 |
+| input/placeholder      | ![#888ca6](https://via.placeholder.com/10/888ca6?text=+) #888ca6 |
+| input/invalid-outline  | ![#ff5353](https://via.placeholder.com/10/ff5353?text=+) #ff5353 |
+| input/selected-outline | ![#091030](https://via.placeholder.com/10/091030?text=+) #091030 |
+
+## 📃 Tipografia
+
+| Tipo  | Especificações  |
+| ----- | --------------- |
+| texto | `14px` `Roboto` |
+
+## 🗣️ Suporte
+
+Para obter ajuda ou mandar um feedback para nossa equipe, envie um email para <labyes23@gmail.com>.
+
+## 👥 Autores
+
+- [Caio Balduino](https://github.com/Caiobaldur) - Design e padronização do layout.
+- [Cristian Sbardelotto](https://github.com/cristian-sbardelotto) - Desenvolvedor Frontend.
+- [Evandro Damaso](https://github.com/dam450) - Design e padronização do layout.
+- [Francisco Ylderlan](https://github.com/FranciscoYlderlan) - Design e padronização do layout.
+- [Grace Kelly](https://github.com/gksouza) - Q.A.
+- [Mateus Villa Nova](https://github.com/Villa-Nova) - Design e padronização do layout.
+- [Thalyta Rangel](https://github.com/ThalytaRangel) - Design e padronização do layout.
