+++ src/presentation/pages/Properties/components/Modal/styles.js
@@ -1,45 +1,9 @@
 import styled, { keyframes } from 'styled-components'
 
-import { PhoneInput } from 'react-international-phone'
-import 'react-international-phone/style.css'
-
 const appear = keyframes`
   from {
     transform: translateY(-10vh);
     opacity: 0;
-  }
-`
-
-export const StyledPhoneInput = styled(PhoneInput)`
-  border: 1px solid
-    ${({ theme, $isValid }) =>
-      $isValid ? theme.COLORS['blue-sky-600'] : theme.COLORS.tomate};
-  color: ${({ theme }) => theme.COLORS['blue-sky-700']};
-
-  border-radius: 5px;
-
-  font-family: Roboto;
-  font-size: 1.4rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 100%;
-
-  .phone-number-input {
-    width: 100%;
-    background-color: ${({ theme }) => theme.COLORS['zinc-50']};
-  }
-
-  &::placeholder {
-    color: ${({ theme }) => theme.COLORS['zinc-200']};
-  }
-
-  &:focus {
-    outline: 1px solid ${({ theme }) => theme.COLORS['blue-sky-500']};
-  }
-
-  &:disabled {
-    filter: contrast(0.75);
-    cursor: not-allowed;
   }
 `
